+++ core/unittest/processor/ProcessorParseTimestampNativeUnittest.cpp
@@ -18,6 +18,12 @@
 #include "config/Config.h"
 #include "processor/ProcessorParseTimestampNative.h"
 #include "plugin/instance/ProcessorInstance.h"
+#include "common/TimeUtil.h"
+#include <string>
+#include <vector>
+
+using namespace std;
+using namespace logtail;
 
 namespace logtail {
 
@@ -36,6 +42,7 @@ public:
     void TestProcessRegularFormatFailed();
     void TestProcessHistoryDiscard();
     void TestProcessEventPreciseTimestampLegacy();
+    void TestCheckTime();
 
     PipelineContext mContext;
 };
@@ -46,6 +53,242 @@ UNIT_TEST_CASE(ProcessorParseTimestampNativeUnittest, TestProcessRegularFormat);
 UNIT_TEST_CASE(ProcessorParseTimestampNativeUnittest, TestProcessRegularFormatFailed);
 UNIT_TEST_CASE(ProcessorParseTimestampNativeUnittest, TestProcessHistoryDiscard);
 UNIT_TEST_CASE(ProcessorParseTimestampNativeUnittest, TestProcessEventPreciseTimestampLegacy);
+UNIT_TEST_CASE(ProcessorParseTimestampNativeUnittest, TestCheckTime);
+
+bool CheckTimeFormatV2(const std::string& timeValue, const std::string& timeFormat) {
+    LogtailTime logTime = {0, 0};
+
+    int nanosecondLength = -1;
+    const char* strptimeResult = NULL;
+    strptimeResult = Strptime(timeValue.c_str(), timeFormat.c_str(), &logTime, nanosecondLength, -1);
+    if (NULL == strptimeResult) {
+        return false;
+    }
+    return true;
+}
+
+bool CheckTimeFormatV1(const std::string& timeValue, const std::string& timeFormat) {
+    struct tm tm;
+
+    if (NULL == strptime(timeValue.c_str(), timeFormat.c_str(), &tm)) {
+        return false;
+    } else {
+        return true;
+    }
+}
+
+void ProcessorParseTimestampNativeUnittest::TestCheckTime() {
+    string timeValue;
+    string timeFormat;
+    timeValue = "Fri";
+    timeFormat = "%a";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "Friday";
+    timeFormat = "%A";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "Jan";
+    timeFormat = "%b";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "January";
+    timeFormat = "%B";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "19";
+    timeFormat = "%d";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "Jan";
+    timeFormat = "%h";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "22";
+    timeFormat = "%H";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "01";
+    timeFormat = "%I";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "08";
+    timeFormat = "%m";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "01";
+    timeFormat = "%M";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "\n";
+    timeFormat = "%n";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "AM";
+    timeFormat = "%p";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "11:59:59 AM";
+    timeFormat = "%r";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "23:59";
+    timeFormat = "%R";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "59";
+    timeFormat = "%S";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = " ";
+    timeFormat = "%t";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "98";
+    timeFormat = "%y";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "2004";
+    timeFormat = "%Y";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "20";
+    timeFormat = "%C";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "31";
+    timeFormat = "%e";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "365";
+    timeFormat = "%j";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "2";
+    timeFormat = "%u";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "53";
+    timeFormat = "%U";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "24";
+    timeFormat = "%V";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "5";
+    timeFormat = "%w";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "23";
+    timeFormat = "%W";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeValue = "Tue Nov 20 14:12:58 2020";
+    timeFormat = "%c";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeFormat = "%x";
+    timeValue = "10/26/23";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeFormat = "%X";
+    timeValue = "14:12:58";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeFormat = "%s";
+    timeValue = "1605853978";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%f";
+    timeValue = "123456789";
+    APSARA_TEST_TRUE_FATAL(!CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%Y-%m-%d %H:%M:%S.%f";
+    timeValue = "2021-11-25 14:16:46.123456789";
+    APSARA_TEST_TRUE_FATAL(!CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%Y-%m-%d %H:%M:%S";
+    timeValue = "2020-11-20 14:12:58";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "[%Y-%m-%d %H:%M:%S";
+    timeValue = "[2017-12-11 15:05:07.012]";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%d %b %y %H:%M";
+    timeValue = "02 Jan 06 15:04 MST";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%d %b %y %H:%M";
+    timeValue = "02 Jan 06 15:04 -0700";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%A, %d-%b-%y %H:%M:%S";
+    timeValue = "Monday, 02-Jan-06 15:04:05 MST";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%A, %d %b %Y %H:%M:%S";
+    timeValue = "Mon, 02 Jan 2006 15:04:05 MST";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%Y-%m-%dT%H:%M:%S";
+    timeValue = "2006-01-02T15:04:05Z07:00";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%Y-%m-%dT%H:%M:%S";
+    timeValue = "2006-01-02T15:04:05.999999999Z07:00";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%s";
+    timeValue = "1637843406";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%s";
+    timeValue = "1637843406123";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%D";
+    timeValue = "11/20/20";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeFormat = "%F";
+    timeValue = "2020-11-20";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+    timeFormat = "%T";
+    timeValue = "14:12:58";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%z";
+    timeValue = "+0800";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%Z";
+    timeValue = "CST";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+
+    timeFormat = "%%";
+    timeValue = "%";
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV1(timeValue, timeFormat));
+    APSARA_TEST_TRUE_FATAL(CheckTimeFormatV2(timeValue, timeFormat));
+}
 
 void ProcessorParseTimestampNativeUnittest::TestInit() {
     Config config;