+++ CHANGELOG.md
@@ -39,6 +39,7 @@ your changes, such as:
 
 - [public] [both] [added] add UsingOldContentTag. When UsingOldContentTag is set to false, the Tag is now placed in the Meta instead of Logs during cgo.
 - [public] [both] [fixed] fix send local buffer failed when upgrade iLogtail from version earlier than 1.3.
+- [public] [both] [updated] Updated strptime_ns to parse %c format from "%x %X" to "%a %b %d %H:%M:%S %Y" for consistent behavior with striptime.
 - [public] [both] [fixed] fix topic key does not support underscore.
 - [public] [both] [fixed] fix jmxfetch status error when exist multi jmxfetch config in the same machine.
 - [public] [both] [fixed] fix user container wss increase problem.